Fix for removing misconfigured subs/csv

+++ 94-delete-operators.sh
@@ -11,10 +11,14 @@ oc get smcp basic -n istio-system 2>/dev/null && echo "Remove the Service Mesh C
 for sub in jaeger-product kiali-ossm servicemeshoperator
 do
 	# Note that if ACM is installed there can be a conflict between "subcriptions" and "subs" which are different APIs!
+	oc get subs $sub -n openshift-operators 
 	CSV=`oc get subs $sub -n openshift-operators -o json | jq -r .status.currentCSV`	
 	oc delete subs $sub -n openshift-operators
-	oc delete csv $CSV -n openshift-operators
+	[ "$CSV" ] && oc delete csv $CSV -n openshift-operators || echo "No csv object found in: subs $sub -n openshift-operators"
 done
+
+# Hammer method!  Sometimes "status.currentCSV" does not exist due to misconfiguration. 
+oc get csv -n openshift-operators -oname | cut -d/ -f2 | grep -e ^kiali-operator -e ^servicemeshoperator -e ^jaeger-operator | xargs -L1 oc delete -n openshift-operators csv || true
 
 echo 
 echo To remove all leftover CRDs, run bin/cleanup-leftover-crds.sh
